+++ src/transport/unicast/link/tx.c
@@ -84,8 +84,8 @@ int8_t _z_unicast_send_z_msg(_z_session_t *zn, _z_zenoh_message_t *z_msg, z_reli
 #endif  // Z_MULTI_THREAD == 1
     } else {
 #if Z_MULTI_THREAD == 1
-        int locked = _z_mutex_trylock(&ztu->_mutex_tx);
-        if (locked != 0) {
+        int8_t locked = _z_mutex_trylock(&ztu->_mutex_tx);
+        if (locked != (int8_t)0) {
             _Z_INFO("Dropping zenoh message because of congestion control\n");
             // We failed to acquire the lock, drop the message
             drop = true;
